Move BaseMessageEventUtil to trace.internal. Work around findbugs.
