+++ api/src/main/java/io/opencensus/trace/export/SpanData.java
@@ -22,13 +22,13 @@ import com.google.auto.value.AutoValue;
 import io.opencensus.common.Timestamp;
 import io.opencensus.trace.Annotation;
 import io.opencensus.trace.AttributeValue;
-import io.opencensus.trace.BaseMessageEventUtil;
 import io.opencensus.trace.Link;
 import io.opencensus.trace.MessageEvent;
 import io.opencensus.trace.Span;
 import io.opencensus.trace.SpanContext;
 import io.opencensus.trace.SpanId;
 import io.opencensus.trace.Status;
+import io.opencensus.trace.internal.BaseMessageEventUtil;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
