+++ api/src/main/java/io/opencensus/trace/internal/BaseMessageEventUtil.java
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-package io.opencensus.trace;
+package io.opencensus.trace.internal;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import io.opencensus.common.Internal;
 
 /**
- * Helper class to convert/cast between for {@link MessageEvent} and {@link NetworkEvent}.
+ * Helper class to convert/cast between for {@link io.opencensus.trace.MessageEvent} and {@link
+ * io.opencensus.trace.NetworkEvent}.
  *
  * @since 0.12
  */
@@ -29,7 +30,8 @@ import io.opencensus.common.Internal;
 @SuppressWarnings("deprecation")
 public final class BaseMessageEventUtil {
   /**
-   * Cast or convert a {@link io.opencensus.trace.BaseMessageEvent} to {@link MessageEvent}.
+   * Cast or convert a {@link io.opencensus.trace.BaseMessageEvent} to {@link
+   * io.opencensus.trace.MessageEvent}.
    *
    * <p>Warning: if the input is a {@code io.opencensus.trace.NetworkEvent} and contains {@code
    * kernelTimestamp} information, this information will be dropped.
@@ -37,17 +39,21 @@ public final class BaseMessageEventUtil {
    * @param event the {@code BaseMessageEvent} that is being cast or converted.
    * @return a {@code MessageEvent} representation of the input.
    */
-  public static MessageEvent asMessageEvent(io.opencensus.trace.BaseMessageEvent event) {
+  public static io.opencensus.trace.MessageEvent asMessageEvent(
+      io.opencensus.trace.BaseMessageEvent event) {
     checkNotNull(event);
-    if (event instanceof MessageEvent) {
-      return (MessageEvent) event;
+    if (event instanceof io.opencensus.trace.MessageEvent) {
+      return (io.opencensus.trace.MessageEvent) event;
+    }
+    if (!(event instanceof io.opencensus.trace.NetworkEvent)) { // Work around FindBugs.
+      throw new IllegalArgumentException("NetworkEvent expected.");
     }
     io.opencensus.trace.NetworkEvent networkEvent = (io.opencensus.trace.NetworkEvent) event;
-    MessageEvent.Type type =
+    io.opencensus.trace.MessageEvent.Type type =
         (networkEvent.getType() == io.opencensus.trace.NetworkEvent.Type.RECV)
-            ? MessageEvent.Type.RECEIVED
-            : MessageEvent.Type.SENT;
-    return MessageEvent.builder(type, networkEvent.getMessageId())
+            ? io.opencensus.trace.MessageEvent.Type.RECEIVED
+            : io.opencensus.trace.MessageEvent.Type.SENT;
+    return io.opencensus.trace.MessageEvent.builder(type, networkEvent.getMessageId())
         .setUncompressedMessageSize(networkEvent.getUncompressedMessageSize())
         .setCompressedMessageSize(networkEvent.getCompressedMessageSize())
         .build();
@@ -66,9 +72,12 @@ public final class BaseMessageEventUtil {
     if (event instanceof io.opencensus.trace.NetworkEvent) {
       return (io.opencensus.trace.NetworkEvent) event;
     }
-    MessageEvent messageEvent = (MessageEvent) event;
+    if (!(event instanceof io.opencensus.trace.MessageEvent)) { // Work around FindBugs.
+      throw new IllegalArgumentException("MessageEvent expected.");
+    }
+    io.opencensus.trace.MessageEvent messageEvent = (io.opencensus.trace.MessageEvent) event;
     io.opencensus.trace.NetworkEvent.Type type =
-        (messageEvent.getType() == MessageEvent.Type.RECEIVED)
+        (messageEvent.getType() == io.opencensus.trace.MessageEvent.Type.RECEIVED)
             ? io.opencensus.trace.NetworkEvent.Type.RECV
             : io.opencensus.trace.NetworkEvent.Type.SENT;
     return io.opencensus.trace.NetworkEvent.builder(type, messageEvent.getMessageId())
